% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goFit.R
\name{goFit}
\alias{goFit}
\title{Measures of Fit for Categorical Response Model}
\usage{
goFit(obj, crit, display, ...)
}
\arguments{
\item{obj}{single, multiple or list of model objects}

\item{crit}{information criterion type on which model comparison is based.}

\item{display}{chooses to display either 'best', 'top' or 'all' result}

\item{...}{.}
}
\value{
\item{Categ}{Number of response category}

\item{logLik}{Log-likelihood of fitted model}

\item{Param}{Number of parameters in model including intercept}

\item{AIC}{Akaike Information Criterion}

\item{AICc}{Akaike Information Criterion for small sample}

\item{BIC}{Bayesian Information Criterion}

\item{AIC.diff}{AIC difference, same also for AICc.diff and BIC.diff}

\item{Akaike.wt}{AIC weight of fitted model}

\item{Schwarz.wt}{BIC weight of fitted model}

\item{ER}{Evidence ratio of fitted model}

\item{Pseudo.R2}{Pseudo R2 of fitted model}
}
\description{
this function performs the goodness-of-fit of categorical outcome models, it returns valuable information measures and pseudo R2s of fitted models.
}
\details{
\code{goFit} supports objects of class glm, vglm, clm2, polr and multinom. In addition to computing several pseudo-R2 measures relevant to the analysis of categorical outcome models, \code{goFit} also conducts model comparison given two or more models in the function. It provides an automatically ranked information criterion table with the best model on top. The initial object names supplied to the function are replaced with an ordered internally generated numbering as follows: (fit1, fit2, ...) with respect to the ordering of objects in the function.
}
\examples{

require("MASS") ## for polr method
set.seed(63)

x1  <- rnorm(200,0,1); x2 <- rnorm(200,2,3)
x3  <- rnorm(200,5,2); x4 <- runif(200,2,5)
cnt <- 0.2*x1 - 0.5*x2 + 0.4*x3 - 0.3*x4 + 2 + rnorm(200, 0, 1)

pdv <- as.factor(cut(cnt, breaks=quantile(cnt), include.lowest=TRUE,
                     labels=c(1, 2, 3, 4)))
qnt <- quantile(cnt)
bdv <- cut(cnt, breaks = c(-Inf,qnt[3],Inf), include.lowest=TRUE, labels = c(0,1))

### Binary & polytmous outcome models

m1 <-  glm(bdv ~ x1 + x2, family = binomial(link = "probit"))
m2 <- polr(pdv ~ x1 * x2 + x3, method = c("logistic"))
m3 <- polr(pdv ~ x1 + x2 * x3, method = c("logistic"))
m4 <- polr(pdv ~ x1 + x2 + x3, method = c("logistic"))
m5 <- polr(pdv ~ x1 * x2 * x3, method = c("logistic"))
m6 <- polr(pdv ~ x1 * x2 + x3 * x4, method = c("logistic"))

mlst <- list(m2, m3, m4, m5, m6)

### single object
goFit(m1)

### multiple object
goFit(m2,m3,m4,crit="AICc",display="top")

### list of objects
goFit(mlst, crit = "BIC",display="all")
}
\references{
Long, J.S. (1997). \emph{Regression Models for Categorical and Limited Dependent Variables}. California: Sage Publications.

Burnham, K. P., & Anderson, D. R. (2002). \emph{Model Selection and Multimodel Inference: A Practical Information-Theoretical Approach}. Springer.

Ugba, E. R. & Gertheiss, J. (2017). A Generalized Likelihood Ratio Index for Discrete Limited Dependent Variable Models — with Emphasis on the Cumulative Link Model. (preprint)
}
\seealso{
\code{\link{plotgoF}}
}
\author{
Ejike R. Ugba
}
