% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgoF.R
\name{plotgoF}
\alias{plotgoF}
\title{Plot Function for Objects of Class goF}
\usage{
plotgoF(obj, type)
}
\arguments{
\item{obj}{saved object of class goF}

\item{type}{respectively selects either 'ps' or "wt" for the pseudo R2 and the information weight.}
}
\description{
A Plot function for goodness of fit measures, it returns plots of pseudo R2's & information weights of categorical response models.
}
\examples{

require("ordinal")
set.seed(63)

x1  <- rnorm(200,0,1); x2 <- rnorm(200,2,3)
x3  <- rnorm(200,5,2); x4 <- runif(200,2,5)
cnt <- 0.2*x1 - 0.5*x2 + 0.4*x3 - 0.3*x4 + 2 + rnorm(200, 0, 1)

qnt <- quantile(cnt)
bdv <- cut(cnt, breaks = c(-Inf,qnt[3],Inf), include.lowest=TRUE,labels = c(0,1))
pdv <- as.factor(cut(cnt, breaks=quantile(cnt), include.lowest=TRUE,
                 labels=c(1, 2, 3, 4), ordered=TRUE))

q1 <- clm2(pdv ~ x1 + x3, link = "cloglog")
q2 <- clm2(pdv ~ x1 + x2 + x4, link = "cauchit")
q3 <- clm2(pdv ~ x1 + x2, link = "logistic")
q4 <- clm2(pdv ~ x1 + x2 + x4, link = "probit")
q5 <- clm2(pdv ~ x1 + x2, link = "logistic")
qlst <- list(q1,q2,q3,q4,q5)

## AIC weights
jr <- goFit(qlst, crit = "AIC", display = "all")
plotgoF(jr, type = "wt")

## Pseudo R2
kr <- glm(bdv ~ x1 + x2, family = binomial())
sr <- goFit(kr, crit = "AICc", display = "all")
plotgoF(sr)
}
\seealso{
\code{\link{goFit}}
}
\author{
Ejike R. Ugba
}
